apps/api/pyproject.toml
[project]
name = "gluex-benchmarking-service"
version = "0.0.1"
description = ""
authors = []
license = { text = "MIT" }
requires-python = ">=3.11,<4.0"


[tool.poetry]
packages = [
    { include = "src", from = "." },
    { include = "scripts", from = "." },
]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
SQLAlchemy = "*"
fastapi = "*"
python-dotenv = "*"
requests = "*"
uvicorn = { version = "*", extras = ["standard"] }
mangum = ">=0.19.0,<0.20.0"
pydantic = "<2.0.0"
psycopg2-binary = "^2.9.10"


[tool.poetry.scripts]
benchmark-api = "src.main:app"
benchmark-run = "scripts.run_automated_benchmark:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
